# lib/api-zod/src/generated/types/accessibilityInfo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * StadiumIQ AI - FIFA World Cup 2026 Stadium Operations Platform
 * OpenAPI spec version: 0.1.0
 */

export interface AccessibilityInfo {
  wheelchairEntrances: string[];
  accessibleRestrooms: string[];
  signLanguageServices: boolean;
  audioDescriptionAvailable: boolean;
  mobilityAidRental: boolean;
  dedicatedVolunteers: number;
  emergencyEvacuationPoints: string[];
}