# lib/api-zod/src/generated/types/announcementResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * StadiumIQ AI - FIFA World Cup 2026 Stadium Operations Platform
 * OpenAPI spec version: 0.1.0
 */

export interface AnnouncementResponse {
  english: string;
  spanish: string;
  french: string;
  portuguese: string;
}