# lib/api-zod/src/generated/types/geminiConversationInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * StadiumIQ AI - FIFA World Cup 2026 Stadium Operations Platform
 * OpenAPI spec version: 0.1.0
 */

export interface GeminiConversationInput {
  title: string;
}