# lib/api-zod/src/generated/types/incident.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * StadiumIQ AI - FIFA World Cup 2026 Stadium Operations Platform
 * OpenAPI spec version: 0.1.0
 */
import type { IncidentSeverity } from './incidentSeverity';
import type { IncidentStatus } from './incidentStatus';
import type { IncidentType } from './incidentType';

export interface Incident {
  id: string;
  type: IncidentType;
  severity: IncidentSeverity;
  location: string;
  description: string;
  status: IncidentStatus;
  reportedAt: Date;
  volunteersAssigned: number;
}