# lib/api-zod/src/generated/types/matchStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * StadiumIQ AI - FIFA World Cup 2026 Stadium Operations Platform
 * OpenAPI spec version: 0.1.0
 */
import type { MatchStatusPhase } from './matchStatusPhase';

export interface MatchStatus {
  homeTeam: string;
  awayTeam: string;
  homeScore: number;
  awayScore: number;
  minute: number;
  phase: MatchStatusPhase;
  venue: string;
}