# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * StadiumIQ AI - FIFA World Cup 2026 Stadium Operations Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StadiumStatusOverallCrowdLevel = typeof StadiumStatusOverallCrowdLevel[keyof typeof StadiumStatusOverallCrowdLevel];


export const StadiumStatusOverallCrowdLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
  critical: 'critical',
} as const;

export type MatchStatusPhase = typeof MatchStatusPhase[keyof typeof MatchStatusPhase];


export const MatchStatusPhase = {
  pre_match: 'pre_match',
  first_half: 'first_half',
  half_time: 'half_time',
  second_half: 'second_half',
  extra_time: 'extra_time',
  full_time: 'full_time',
} as const;

export interface MatchStatus {
  homeTeam: string;
  awayTeam: string;
  homeScore: number;
  awayScore: number;
  minute: number;
  phase: MatchStatusPhase;
  venue: string;
}

export interface Weather {
  condition: string;
  temperatureCelsius: number;
  humidity: number;
  windKph: number;
  uvIndex: number;
}

export interface StadiumStatus {
  matchStatus: MatchStatus;
  weather: Weather;
  overallCrowdLevel: StadiumStatusOverallCrowdLevel;
  totalAttendance: number;
  capacityPercent: number;
  updatedAt: string;
}

export type GateStatus = typeof GateStatus[keyof typeof GateStatus];


export const GateStatus = {
  open: 'open',
  congested: 'congested',
  closed: 'closed',
  restricted: 'restricted',
} as const;

export interface Gate {
  id: string;
  name: string;
  status: GateStatus;
  densityPercent: number;
  queueLengthMinutes: number;
  sector: string;
  accessibilityEnabled: boolean;
}

export type CrowdZoneLevel = typeof CrowdZoneLevel[keyof typeof CrowdZoneLevel];


export const CrowdZoneLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
  critical: 'critical',
} as const;

export interface CrowdZone {
  zoneId: string;
  zoneName: string;
  densityPercent: number;
  level: CrowdZoneLevel;
  capacity: number;
  current: number;
}

export type TransportRouteType = typeof TransportRouteType[keyof typeof TransportRouteType];


export const TransportRouteType = {
  metro: 'metro',
  bus: 'bus',
  shuttle: 'shuttle',
  taxi: 'taxi',
  walking: 'walking',
} as const;

export type TransportRouteStatus = typeof TransportRouteStatus[keyof typeof TransportRouteStatus];


export const TransportRouteStatus = {
  on_time: 'on_time',
  delayed: 'delayed',
  disrupted: 'disrupted',
  suspended: 'suspended',
} as const;

export type TransportRouteCrowdLevel = typeof TransportRouteCrowdLevel[keyof typeof TransportRouteCrowdLevel];


export const TransportRouteCrowdLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
} as const;

export interface TransportRoute {
  id: string;
  type: TransportRouteType;
  name: string;
  status: TransportRouteStatus;
  delayMinutes: number;
  nextDeparture: string;
  destination: string;
  crowdLevel: TransportRouteCrowdLevel;
}

export type IncidentType = typeof IncidentType[keyof typeof IncidentType];


export const IncidentType = {
  medical: 'medical',
  security: 'security',
  crowd_surge: 'crowd_surge',
  infrastructure: 'infrastructure',
  weather: 'weather',
  transport: 'transport',
} as const;

export type IncidentSeverity = typeof IncidentSeverity[keyof typeof IncidentSeverity];


export const IncidentSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export type IncidentStatus = typeof IncidentStatus[keyof typeof IncidentStatus];


export const IncidentStatus = {
  active: 'active',
  resolving: 'resolving',
  resolved: 'resolved',
} as const;

export interface Incident {
  id: string;
  type: IncidentType;
  severity: IncidentSeverity;
  location: string;
  description: string;
  status: IncidentStatus;
  reportedAt: string;
  volunteersAssigned: number;
}

export type VolunteerRole = typeof VolunteerRole[keyof typeof VolunteerRole];


export const VolunteerRole = {
  gate_marshal: 'gate_marshal',
  crowd_guide: 'crowd_guide',
  medical: 'medical',
  transport: 'transport',
  accessibility: 'accessibility',
  general: 'general',
} as const;

export type VolunteerStatus = typeof VolunteerStatus[keyof typeof VolunteerStatus];


export const VolunteerStatus = {
  available: 'available',
  assigned: 'assigned',
  on_break: 'on_break',
  off_duty: 'off_duty',
} as const;

export interface Volunteer {
  id: string;
  name: string;
  role: VolunteerRole;
  zone: string;
  status: VolunteerStatus;
  languages: string[];
  /** @nullable */
  currentTask: string | null;
}

export type AlertLevel = typeof AlertLevel[keyof typeof AlertLevel];


export const AlertLevel = {
  info: 'info',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface Alert {
  id: string;
  level: AlertLevel;
  title: string;
  message: string;
  /** @nullable */
  zone: string | null;
  createdAt: string;
  acknowledged: boolean;
}

export interface AccessibilityInfo {
  wheelchairEntrances: string[];
  accessibleRestrooms: string[];
  signLanguageServices: boolean;
  audioDescriptionAvailable: boolean;
  mobilityAidRental: boolean;
  dedicatedVolunteers: number;
  emergencyEvacuationPoints: string[];
}

export type AssistRequestPersona = typeof AssistRequestPersona[keyof typeof AssistRequestPersona];


export const AssistRequestPersona = {
  fan: 'fan',
  operations: 'operations',
  volunteer: 'volunteer',
} as const;

export interface AssistRequest {
  query: string;
  persona: AssistRequestPersona;
  language?: string;
}

export type AssistResponseUrgency = typeof AssistResponseUrgency[keyof typeof AssistResponseUrgency];


export const AssistResponseUrgency = {
  normal: 'normal',
  elevated: 'elevated',
  urgent: 'urgent',
} as const;

export interface AssistResponse {
  response: string;
  suggestions: string[];
  urgency: AssistResponseUrgency;
  relatedZones: string[];
}

export type OperationalRecommendationPriority = typeof OperationalRecommendationPriority[keyof typeof OperationalRecommendationPriority];


export const OperationalRecommendationPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export type OperationalRecommendationCategory = typeof OperationalRecommendationCategory[keyof typeof OperationalRecommendationCategory];


export const OperationalRecommendationCategory = {
  crowd_management: 'crowd_management',
  gate_control: 'gate_control',
  transport: 'transport',
  volunteer: 'volunteer',
  safety: 'safety',
  accessibility: 'accessibility',
} as const;

export interface OperationalRecommendation {
  id: string;
  priority: OperationalRecommendationPriority;
  category: OperationalRecommendationCategory;
  title: string;
  recommendation: string;
  affectedZones: string[];
  generatedAt: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
}

export interface GeminiMessageInput {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface ApiError {
  error: string;
}

export interface AnnouncementRequest {
  incidentId: string;
  incidentDescription: string;
  severity: string;
  zone?: string;
}

export interface AnnouncementResponse {
  english: string;
  spanish: string;
  french: string;
  portuguese: string;
}